refactor the way the config is used
