+++ config/config_test.go
@@ -1,4 +1,4 @@
-// Copyright 2019 The Prometheus Authors
+// Copyright 2020 The Prometheus Authors
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package langserver
+package config
 
 import (
 	"os"
@@ -36,13 +36,11 @@ func TestUnmarshalENV(t *testing.T) {
 		{
 			title: "full config",
 			variables: map[string]string{
-				"LANGSERVER_RPCTRACE":      "text",
 				"LANGSERVER_PROMETHEUSURL": "http://localhost:9090",
 				"LANGSERVER_RESTAPIPORT":   "8080",
 				"LANGSERVER_LOGFORMAT":     "json",
 			},
 			expected: &Config{
-				RPCTrace:      "text",
 				PrometheusURL: "http://localhost:9090",
 				RESTAPIPort:   8080,
 				LogFormat:     JSONFormat,
