+++ langserver/server.go
@@ -28,6 +28,7 @@ import (
 	"os"
 	"sync"
 
+	"github.com/prometheus-community/promql-langserver/config"
 	promClient "github.com/prometheus-community/promql-langserver/prometheus"
 
 	"github.com/go-kit/kit/log"
@@ -57,8 +58,6 @@ type server struct {
 
 	cache cache.DocumentCache
 
-	config *Config
-
 	metadataService promClient.MetadataService
 
 	lifetime context.Context
@@ -87,7 +86,6 @@ func CreateHeadlessServer(ctx context.Context, metadataService promClient.Metada
 	s := &server{
 		client:          &headlessClient{logger: logger},
 		headless:        true,
-		config:          &Config{PrometheusURL: metadataService.GetURL()},
 		metadataService: metadataService,
 	}
 
@@ -105,14 +103,22 @@ func CreateHeadlessServer(ctx context.Context, metadataService promClient.Metada
 }
 
 // ServerFromStream generates a Server from a jsonrpc2.Stream.
-func ServerFromStream(ctx context.Context, stream jsonrpc2.Stream, config *Config) (context.Context, Server) {
+func ServerFromStream(ctx context.Context, stream jsonrpc2.Stream, conf *config.Config) (context.Context, Server) {
 	s := &server{}
 
-	switch config.RPCTrace {
-	case "text":
+	switch conf.LogFormat {
+	case config.TextFormat:
 		stream = protocol.LoggingStream(stream, os.Stderr)
-	case "json":
+	case config.JSONFormat:
 		stream = jSONLogStream(stream, os.Stderr)
+	default:
+		err := fmt.Errorf("invalid log format: '%s'", conf.LogFormat)
+		// nolint: errcheck
+		s.client.ShowMessage(s.lifetime, &protocol.ShowMessageParams{
+			Type:    protocol.Error,
+			Message: err.Error(),
+		})
+		panic(err)
 	}
 
 	s.Conn = jsonrpc2.NewConn(stream)
@@ -120,11 +126,9 @@ func ServerFromStream(ctx context.Context, stream jsonrpc2.Stream, config *Confi
 
 	s.Conn.AddHandler(protocol.ServerHandler(s))
 
-	s.config = config
-
 	s.lifetime, s.exit = context.WithCancel(ctx)
 
-	prometheusClient, err := promClient.NewClient("")
+	prometheusClient, err := promClient.NewClient(conf.PrometheusURL)
 	if err != nil {
 		// nolint: errcheck
 		s.client.ShowMessage(s.lifetime, &protocol.ShowMessageParams{
@@ -138,21 +142,9 @@ func ServerFromStream(ctx context.Context, stream jsonrpc2.Stream, config *Confi
 	return ctx, Server{s}
 }
 
-func (s *server) connectPrometheus(url string) error {
-	if err := s.metadataService.ChangeDataSource(url); err != nil {
-		// nolint: errcheck
-		s.client.ShowMessage(s.lifetime, &protocol.ShowMessageParams{
-			Type:    protocol.Error,
-			Message: fmt.Sprintf("Failed to connect to Prometheus at %s:\n\n%s ", url, err.Error()),
-		})
-		return err
-	}
-	return nil
-}
-
 // RunTCPServer generates a server listening on the provided TCP Address, creating a new language Server
 // instance using plain HTTP for every connection.
-func RunTCPServer(ctx context.Context, addr string, config *Config) error {
+func RunTCPServer(ctx context.Context, addr string, conf *config.Config) error {
 	ln, err := net.Listen("tcp", addr)
 	if err != nil {
 		return err
@@ -164,12 +156,12 @@ func RunTCPServer(ctx context.Context, addr string, config *Config) error {
 			return err
 		}
 
-		go ServerFromStream(ctx, jsonrpc2.NewHeaderStream(conn, conn), config)
+		go ServerFromStream(ctx, jsonrpc2.NewHeaderStream(conn, conn), conf)
 	}
 }
 
 // StdioServer generates a Server instance talking to stdio.
-func StdioServer(ctx context.Context, config *Config) (context.Context, Server) {
+func StdioServer(ctx context.Context, conf *config.Config) (context.Context, Server) {
 	stream := jsonrpc2.NewHeaderStream(os.Stdin, os.Stdout)
-	return ServerFromStream(ctx, stream, config)
+	return ServerFromStream(ctx, stream, conf)
 }
